+++ _tests/cli/options/trade_opt_unit_test.py
@@ -16,7 +16,7 @@ def test__trade_option__returns_the_value__when_cli_is_invoked(
 
 
 def test__trade_wrong_option_value_shows_an_error__when_cli_is_invoked():
-    expected_message = "Only 'BUY', 'SELL' is allowed"
+    expected_message = "Only BUY, SELL is allowed"
 
     result = runner.invoke(cli, ["--trade", "test", "fake-run"])
