+++ src/cli/__init__.py
@@ -1,1 +1,1 @@
-from .options import app as cli, Defaults
+from .options import app as cli
