+++ src/cli/options/__init__.py
@@ -2,14 +2,13 @@ import typing
 
 import typer
 
-from ..defaults import Defaults
-from ..commands import run as run_service
-from .trade import main as __trade_option
-from .version import main as __version_option
 from .country import main as __country_option
 from .fetch_time import main as __set__fetch_time_option
 from .process_time import main as __set__process_time_option
-
+from .trade import main as __trade_option
+from .version import main as __version_option
+from ..commands import run as run_service
+from ..defaults import Defaults
 
 app = typer.Typer()
 
@@ -26,19 +25,19 @@ def fake_run():
 
 @app.callback()
 def main(
-    _version: typing.Optional[bool] = __version_option(typer),
-    _trade: typing.Optional[str] = __trade_option(
-        typer, default=Defaults.trade
-    ),
-    _country: typing.Optional[str] = __country_option(
-        typer, default=Defaults.country
-    ),
-    _set_fetch_time: typing.Optional[int] = __set__fetch_time_option(
-        typer, default=Defaults.set_fetch_time
-    ),
-    _set_process_time: typing.Optional[int] = __set__process_time_option(
-        typer, default=Defaults.set_process_time
-    ),
+        _version: typing.Optional[bool] = __version_option(typer),
+        _trade: typing.Optional[str] = __trade_option(
+            typer, default=Defaults.trade
+        ),
+        _country: typing.Optional[str] = __country_option(
+            typer, default=Defaults.country
+        ),
+        _set_fetch_time: typing.Optional[int] = __set__fetch_time_option(
+            typer, default=Defaults.set_fetch_time
+        ),
+        _set_process_time: typing.Optional[int] = __set__process_time_option(
+            typer, default=Defaults.set_process_time
+        ),
 ) -> None:
     Defaults.trade = _trade
     Defaults.country = _country
